refactor navigation factory and add config types
