+++ src/commands/index.ts
@@ -1,1 +1,1 @@
-export {};
+export type { Command, CliCommand, VoiceCommand } from '../interfaces/commands';
