+++ src/interfaces/config.ts
@@ -1,0 +1,8 @@
+export interface SupabaseConfig {
+  url: string;
+  anonKey: string;
+}
+
+export interface AnalyticsConfig {
+  disabled?: boolean;
+}
