+++ src/interfaces/index.ts
@@ -8,3 +8,4 @@ export type { Network } from './network';
 export type { Notifications, PhaseEmitter } from './notifications';
 export type { PhaseSync, PhaseRecord } from './phaseSync';
 export type { SupabaseService } from './supabaseService';
+export type { SupabaseConfig, AnalyticsConfig } from './config';
