+++ src/processors/grouped/index.ts
@@ -1,1 +1,1 @@
-export {};
+export type { GroupedProcessor } from '../../interfaces/processors/grouped';
