+++ src/processors/index.ts
@@ -1,1 +1,1 @@
-export {};
+export type { Processor } from '../interfaces/processors';
