+++ src/services/analytics.ts
@@ -1,15 +1,22 @@
 import firebaseAnalytics from '@react-native-firebase/analytics';
 import { log } from './logger';
 import type { AnalyticsService } from '../interfaces/analyticsService';
+import type { AnalyticsConfig } from '../interfaces/config';
 
-export const analytics: AnalyticsService = {
-  async trackEvent(name, params) {
-    try {
-      await firebaseAnalytics().logEvent(name, params);
-    } catch (err) {
-      await log('WARN', `Failed to log analytics event: ${String(err)}`);
-    }
-  },
-};
+export function createAnalytics(
+  config: AnalyticsConfig = {},
+): AnalyticsService {
+  return {
+    async trackEvent(name, params) {
+      if (config.disabled) return;
+      try {
+        await firebaseAnalytics().logEvent(name, params);
+      } catch (err) {
+        await log('WARN', `Failed to log analytics event: ${String(err)}`);
+      }
+    },
+  };
+}
 
+export const analytics = createAnalytics();
 export const trackEvent = analytics.trackEvent;
