+++ src/services/supabase.ts
@@ -8,17 +8,20 @@ import { network } from './network';
 import { log } from './logger';
 import type { Light, LightCycle } from '../domain/types';
 import type { SupabaseService } from '../interfaces/supabaseService';
+import type { SupabaseConfig } from '../interfaces/config';
 
-const SUPABASE_URL = process.env.EXPO_PUBLIC_SUPABASE_URL ?? '';
-const SUPABASE_ANON_KEY = process.env.EXPO_PUBLIC_SUPABASE_ANON_KEY ?? '';
-
-export const supabase: SupabaseClient = createClient(
-  SUPABASE_URL,
-  SUPABASE_ANON_KEY,
-  {
+export function createSupabaseClient(config: SupabaseConfig): SupabaseClient {
+  return createClient(config.url, config.anonKey, {
     global: { fetch: network.fetchWithTimeout },
-  },
-);
+  });
+}
+
+const defaultConfig: SupabaseConfig = {
+  url: process.env.EXPO_PUBLIC_SUPABASE_URL ?? '',
+  anonKey: process.env.EXPO_PUBLIC_SUPABASE_ANON_KEY ?? '',
+};
+
+export const supabase = createSupabaseClient(defaultConfig);
 
 async function fetchLightsAndCycles(): Promise<{
   lights: Light[];
