+++ src/sources/index.ts
@@ -1,1 +1,1 @@
-export {};
+export type { Source } from '../interfaces/sources';
