+++ src/stores/index.ts
@@ -1,1 +1,1 @@
-export {};
+export type { Store } from '../interfaces/stores';
